[refactor] useThreadListQuery thread 폴더로 이동
